requirements.txt
easydict==1.13
matplotlib==3.10.8
numpy==2.4.2
opencv_python==4.12.0.88
Pillow==12.1.0
PyYAML==6.0.3
torch==2.8.0+cu128
torchvision==0.23.0+cu128
tqdm==4.66.2
transformers==4.57.3
